requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.115.6
h11==0.14.0
httptools==0.6.1
idna==3.6
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-experimental>=0.0.40
langchain-pinecone>=0.0.1
langchain-text-splitters>=0.0.1
langchain-upstage>=0.0.1
openai==1.59.7
passlib==1.7.4
Pillow==9.5.0
pinecone-client>=2.2.0
PyMuPDF==1.25.1
PyMySQL==1.1.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart>=0.0.20
PyYAML==6.0.1
requests>=2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.31
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
